linebotapi/requirements.txt
fastapi
jinja2
line-bot-sdk
uvicorn[standard]